requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.6
mysql-connector-python==8.3.0
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
